Add defensive check after filter for cart items
If an extension uses the `woocommerce_get_cart_item_from_session` filter
and modifies the array of cart item data, it can cause a fatal error if
the modified data does not include the specific `data` key containing
an instance of the cart item's corresponding product object. This
simply ensures that even if an extension is using that filter
incorrectly, it will not cause a fatal error.

Fixes #55103