Accurate slug comparison by decoding both sides
This fixes a regression caused by #54940, which was in turn fixing a
regression caused by #51637. When comparing an expected and a specified
product category slug, one or the other may be URL-encoded if it
contains multibyte characters. This simply ensures that both sides have
been decoded before the comparison is made, and avoiding potential
infinite redirect loops.

Note: this is based on #55415 but targeted at the 9.7 release branch
instead of trunk.

Co-authored-by: Pierre <1941656+strategio@users.noreply.github.com>