Add placeholder block for Woo generated content
This commit adds very basic block that act as a placeholder for
the Woo generated content for the transactional emails (order details, shipping address, etc.).

It is registered only in the email editor via JS so that the block is not globally available.
The idea is that we will extend this block to preview Woo Content in the editor per email type.