Use block email for transactional email when defined.
When the feature is active and there is an email created in the block
editor and associated with the transactional email, we render the block email
using MailPoet\EmailEditor\Engine\Renderer\Renderer.
We grab the main content of Woo email and place it instead of the placeholder
generated by Woo Content block.

To grab only the main content, we need to unregister filters for the email header and footer.
We then restore the filters to be able to correctly send other emails that might not have
an associated email created in the block editor..