Moved settings to BISN class.
WC core adds all options with its default value to the database. However, this doesn't work well for merged BIS code where the db option acts as an override for the rollout assignment. This trick relies on the default behaviour of WP settings to use the default value (which can change in code) until the underlying option is set.